consolidate class generation to single module
