+++ lib/phoenix_ui/components/list.ex
@@ -50,37 +50,44 @@ defmodule PhoenixUI.Components.List do
   """
   @spec list_item(Socket.assigns()) :: Rendered.t()
   def list_item(prev_assigns) do
-    assigns =
-      prev_assigns
-      |> assign_class(~w(
+    prev_assigns
+    |> assign_class(~w(
         block w-full whitespace-nowrap py-2 px-4 text-sm font-normal
         text-slate-700 dark:text-slate-300 hover:bg-slate-100 dark:hover:bg-slate-100/25
       ))
-      |> assign_rest([:element])
+    |> assign_rest([:element])
+    |> list_item_markup()
+  end
 
+  defp list_item_markup(%{href: _} = assigns) do
+    ~H"""
+    <.link name={@element} {@rest}>
+      <%= render_slot(@inner_block) %>
+    </.link>
+    """
+  end
+
+  defp list_item_markup(%{navigate: _} = assigns) do
+    ~H"""
+    <.link name={@element} {@rest}>
+      <%= render_slot(@inner_block) %>
+    </.link>
+    """
+  end
+
+  defp list_item_markup(%{patch: _} = assigns) do
+    ~H"""
+    <.link {@rest}>
+      <%= render_slot(@inner_block) %>
+    </.link>
+    """
+  end
+
+  defp list_item_markup(assigns) do
     ~H"""
     <.dynamic_tag name={@element} {@rest}>
       <%= render_slot(@inner_block) %>
     </.dynamic_tag>
     """
-  end
-
-  @doc """
-  Returns all possible component classes for Tailwind CSS JIT compilation.
-
-  ## Examples
-
-      iex> classes()
-      ["class1", "class2", ...]
-
-  """
-  @spec classes :: [String.t()]
-  def classes do
-    [
-      generate_all_classes(&list/1, []),
-      generate_all_classes(&list_item/1, [])
-    ]
-    |> List.flatten()
-    |> Enum.uniq()
   end
 end
