+++ lib/phoenix_ui/components/modal.ex
@@ -58,24 +58,6 @@ defmodule PhoenixUI.Components.Modal do
     """
   end
 
-  @doc """
-  Returns all possible component classes for Tailwind CSS JIT compilation.
-
-  ## Examples
-
-      iex> classes()
-      ["class1", "class2", ...]
-
-  """
-  @spec classes :: [String.t()]
-  def classes do
-    PhoenixUI.Helpers.generate_all_classes(&modal/1,
-      id: ["modal"],
-      max_width: [:xs, :sm, :md, :lg, :xl],
-      open: [true, false]
-    )
-  end
-
   ### JS Interactions ##########################
 
   @doc """
