+++ lib/phoenix_ui/components/typography.ex
@@ -40,24 +40,6 @@ defmodule PhoenixUI.Components.Typography do
     """
   end
 
-  @doc """
-  Returns all possible component classes for Tailwind CSS JIT compilation.
-
-  ## Examples
-
-      iex> classes()
-      ["class1", "class2", ...]
-
-  """
-  @spec classes :: [String.t()]
-  def classes do
-    generate_all_classes(&typography/1,
-      align: ["center", "justify", "left", "right"],
-      color: Theme.colors(),
-      variant: ["h1", "h2", "h3", "h4", "p"]
-    )
-  end
-
   ### CSS Classes ##########################
 
   # Align
